+++ packages/tswebextension/src/lib/common/constants.ts
@@ -1,2 +1,3 @@
 export const USER_FILTER_ID = 0;
+export const ALLOWLIST_FILTER_ID = 100;
 export const BACKGROUND_TAB_ID = -1;