+++ packages/tswebextension/src/lib/mv2/background/api.ts
@@ -3,7 +3,17 @@
  * Needed for backward compatibility during internal API updates.
  * Will be removed in the future.
  */
-import { allowlistApi } from './allowlist';
-import { TabsApi } from './tabs';
+import { Allowlist } from './allowlist';
+import { EngineApi } from './engine-api';
+import { DocumentApi } from './document-api';
+import { TabsApi } from './tabs/tabs-api';
+import { stealthApi } from './stealth-api';
+import { appContext } from './context';
 
-export const tabsApi = new TabsApi(allowlistApi);
+export const allowlist = new Allowlist();
+
+export const engineApi = new EngineApi(allowlist, appContext, stealthApi);
+
+export const documentApi = new DocumentApi(allowlist, engineApi);
+
+export const tabsApi = new TabsApi(documentApi);